Force float32 in certain methods to reduce memory consumption

+++ venomai/preprocess.py
@@ -14,7 +14,7 @@ def rescale_image(image, res, target_res=6, interpolation=cv2.INTER_CUBIC):
 
 def srgb_to_linear(srgb): 
     # https://stackoverflow.com/questions/596216/formula-to-determine-perceived-brightness-of-rgb-color/56678483#56678483
-    srgb = srgb / 255.0
+    srgb = srgb.astype('float32') / 255.0
     
     upper = ((srgb + 0.055) / 1.055)**2.4
     lower = srgb / 12.92
@@ -24,7 +24,7 @@ def srgb_to_linear(srgb):
     return linear
 
 def linear_to_srgb(linear):
-    linear = linear / 255.0
+    linear = linear.astype('float32') / 255.0
     
     upper = 1.055 * linear**(1.0 / 2.4) - 0.055
     lower = 12.92 * linear
